Split monolithic tests
Register each test as its own CTest entry.
